Fix ownership of Path and Buffer, move to FilerUtils.js
